+++ groups/bal/ball/ball_recordbuffer.t.cpp
@@ -343,7 +343,7 @@ void my_DummyDeleter::deleteObject(ball::Record *)
     int my_RecordBuffer::length() const
     {
         bslmt::LockGuard<bslmt::RecursiveMutex> guard(&d_mutex);
-        return d_buffer.size();
+        return static_cast<int>(d_buffer.size());
     }
 //..
 // Note that we always implement a virtual destructor (non-inline) in the .cpp
